Add demo structure for server
