--- packages/server/controllers/topicController.ts
@@ -1,9 +1,0 @@
-import { Request, Response } from 'express'
-
-export const getAllTopics = async (_req: Request, _res: Response) => {
-  throw new Error('Not implemented')
-}
-
-export const createTopic = async (_req: Request, _res: Response) => {
-  throw new Error('Not implemented')
-}
