+++ packages/server/middleware/auth.ts
