+++ packages/server/services/topic.ts
@@ -1,0 +1,12 @@
+const getAllTopics = async () => {
+  return []
+}
+
+const createTopic = async () => {
+  throw new Error('Not implemented')
+}
+
+export const TopicService = {
+  getAllTopics,
+  createTopic,
+}
